requirements.txt
amqp==2.6.0
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Django==3.0.7
django-celery-beat==2.0.0
django-cors-headers==3.4.0
django-rest-knox==4.1.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-gis==0.15
dynaconf==2.2.3
future==0.18.2
idna==2.9
importlib-metadata==1.7.0
kombu==4.6.11
pathspec==0.8.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
python-box==3.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.6.8
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
